Integrated Flask Boilerplate into this repo (https://github.com/taeram/flask-boilerplate)
