+++ app/controllers/gallery.py
@@ -1,7 +1,5 @@
 from app import app
-import os
 from flask import abort, \
-                  flash, \
                   redirect, \
                   render_template, \
                   request, \
@@ -9,23 +7,13 @@ from flask import abort, \
                   url_for
 from flask.ext.login import current_user, \
                             login_user, \
-                            logout_user, \
                             login_required
-from .forms import LoginForm,\
-                   GalleryForm,\
-                   UserForm
-from .database import find_user_by_name, \
-                      find_gallery_all, \
-                      find_gallery_by_id, \
-                      find_gallery_by_share_code, \
-                      find_user_by_id,\
-                      find_user_all,\
-                      db, \
-                      Gallery, \
-                      Photo, \
-                      User
-from .helpers import delete_photo, \
-                     get_gallery_photos
+from app.forms.gallery import GalleryForm
+from app.models.gallery import Gallery
+from app.models.photo import Photo
+from app.models.user import User
+from app.controllers.helpers.photo import delete_photo
+from app.controllers.helpers.gallery import get_gallery_photos
 import json
 import base64
 import hmac
@@ -43,7 +31,7 @@ def home():
     else:
         search_query = ""
 
-    return render_template('index.html', q=search_query)
+    return render_template('gallery/index.html', q=search_query)
 
 
 @app.route('/s/<string:share_code>')
@@ -59,7 +47,7 @@ def share(share_code):
         session.permanent = False
         app.permanent_session_lifetime = timedelta(minutes=1)
 
-    gallery = find_gallery_by_share_code(share_code)
+    gallery = Gallery.find_by_share_code(share_code)
     if not gallery:
         abort(404)
 
@@ -74,36 +62,7 @@ def share(share_code):
     if len(photos) > 0:
         og_photo_url = 'https://%s.s3.amazonaws.com/%s/%s_thumb.jpg' % (app.config['AWS_S3_BUCKET'], gallery.folder, photos[0]['name']);
 
-    return render_template('index.html', json=json.dumps(response), page_title=response['name'], og_photo_url=og_photo_url)
-
-
-@app.route("/login", methods=["GET", "POST"])
-def login():
-    """ Login page """
-    if current_user.is_authenticated():
-        return redirect(url_for('home'))
-
-    form = LoginForm()
-    if form.validate_on_submit():
-        user = find_user_by_name(form.username.data)
-        if user is None or not user.is_valid_password(form.password.data):
-            flash('Invalid username or password', 'danger')
-        elif login_user(user, remember=form.remember.data):
-            # Enable session expiration only if user hasn't chosen to be remembered.
-            session.permanent = not form.remember.data
-            return redirect(request.args.get('next') or url_for('home'))
-    elif form.errors:
-        flash('Invalid username or password', 'danger')
-
-    return render_template('login.html', form=form)
-
-
-@app.route('/logout')
-@login_required
-def logout():
-    """ Logout the user """
-    logout_user()
-    return redirect(url_for('login'))
+    return render_template('gallery/index.html', json=json.dumps(response), page_title=response['name'], og_photo_url=og_photo_url)
 
 
 @app.route('/create/', methods=['GET', 'POST'])
@@ -119,12 +78,11 @@ def gallery_create():
             name=form.name.data,
             created=form.date.data
         )
-        db.session.add(gallery)
-        db.session.commit()
+        gallery.save()
 
         return redirect(url_for('gallery_upload', gallery_id=gallery.id))
 
-    return render_template('gallery.html',
+    return render_template('gallery/update.html',
         form=form,
         page_title="Create an Album",
         form_action=url_for('gallery_create'),
@@ -139,7 +97,7 @@ def gallery_update(gallery_id):
     if current_user.role == "guest":
         abort(404)
 
-    gallery = find_gallery_by_id(gallery_id)
+    gallery = Gallery.find_by_id(gallery_id)
     if not gallery:
         abort(404)
 
@@ -153,147 +111,15 @@ def gallery_update(gallery_id):
         if form.validate_on_submit():
             gallery.name = form.name.data
             gallery.created = form.date.data
-
-            db.session.add(gallery)
-            db.session.commit()
+            gallery.save()
 
             return redirect(url_for('home'))
 
-    return render_template('gallery.html',
+    return render_template('gallery/update.html',
         form=form,
         page_title="Update %s" % gallery.name,
         form_action=url_for('gallery_update', gallery_id=gallery.id),
         form_submit_button_title="Update"
-    )
-
-
-@app.route('/user/list/', methods=['GET'])
-@login_required
-def user_list():
-    """ List all users """
-    if not current_user.role == "admin":
-        abort(404)
-
-    users = find_user_all()
-
-    return render_template('user_list.html',
-        users=users,
-        page_title="Users"
-    )
-
-@app.route('/user/add/', methods=['GET', 'POST'])
-@login_required
-def user_create():
-    """ Create a user """
-    if not current_user.role == "admin":
-        abort(404)
-
-    form = UserForm()
-    if request.method == 'POST' and form.validate_on_submit():
-        user = User(
-            name = form.name.data,
-            password = form.password.data,
-            role = form.role.data
-        )
-        db.session.add(user)
-        db.session.commit()
-
-        return redirect(url_for('user_list'))
-
-    return render_template('user_create.html',
-        form=form,
-        page_title="Create a User",
-        form_action=url_for('user_create'),
-        form_submit_button_title="Create"
-    )
-
-@app.route('/user/update/<int:user_id>', methods=['GET', 'POST'])
-@login_required
-def user_update(user_id):
-    """ Update a user """
-    if not current_user.role == "admin":
-        abort(404)
-
-    user = find_user_by_id(user_id)
-    if not user:
-        abort(404)
-
-    form = UserForm(obj=user)
-    del form.password
-    if request.method == 'POST' and form.validate_on_submit():
-        user.name = form.name.data
-        user.role = form.role.data
-
-        db.session.add(user)
-        db.session.commit()
-
-        return redirect(url_for('user_list'))
-
-    return render_template('user_update.html',
-        user=user,
-        form=form,
-        page_title="Update %s" % user.name ,
-        form_action=url_for('user_update', user_id=user.id),
-        form_submit_button_title="Update"
-    )
-
-@app.route('/user/delete/<int:user_id>', methods=['GET'])
-@login_required
-def user_delete(user_id):
-    """ Delete a user """
-    if not current_user.role == "admin":
-        abort(404)
-
-    user = find_user_by_id(user_id)
-    if not user:
-        abort(404)
-
-    db.session.delete(user)
-    db.session.commit()
-
-    return redirect(url_for('user_list'))
-
-
-@app.route('/change-password/', methods=['GET', 'POST'])
-@login_required
-def user_change_password():
-    """ Change a user's password """
-
-    # Is this an admin resetting a user's password?
-    if request.args.get('user_id'):
-        user_id = request.args.get('user_id')
-
-        if user_id != current_user.id and not current_user.role == "admin":
-            abort(404)
-    else:
-        user_id = current_user.id
-
-    user = find_user_by_id(user_id)
-    if not user:
-        abort(404)
-
-    form = UserForm(obj=user)
-    del form.name
-    del form.role
-    if request.method == 'POST' and form.validate_on_submit():
-        user.set_password(form.password.data)
-
-        db.session.add(user)
-        db.session.commit()
-
-        return redirect(url_for('user_change_password', user_id=user.id))
-
-    if current_user.role == "admin":
-        page_title = "Change password for %s" % user.name
-    else:
-        page_title = "Change your Password"
-
-    return render_template('user_change_password.html',
-        user=user,
-        form=form,
-        page_title=page_title,
-        form_action=url_for('user_change_password', user_id=user.id),
-        form_submit_button_title="Change"
     )
 
 @app.route('/upload/<int:gallery_id>')
@@ -303,7 +129,7 @@ def gallery_upload(gallery_id):
     if current_user.role == "guest":
         abort(404)
 
-    gallery = find_gallery_by_id(gallery_id)
+    gallery = Gallery.find_by_id(gallery_id)
     if not gallery:
         abort(404)
 
@@ -324,7 +150,7 @@ def gallery_upload(gallery_id):
     policy = base64.b64encode(json.dumps(s3_policy))
     signature = base64.b64encode(hmac.new(app.config['AWS_SECRET_ACCESS_KEY'], policy, hashlib.sha1).digest())
 
-    return render_template('upload.html',
+    return render_template('gallery/upload.html',
         gallery=gallery,
         aws_access_key_id=app.config['AWS_ACCESS_KEY_ID'],
         s3_acl=s3_acl,
@@ -344,7 +170,7 @@ def gallery_verify(gallery_id):
     if current_user.role == "guest":
         abort(404)
 
-    return render_template('index.html')
+    return render_template('gallery/index.html')
 
 @app.route('/verify/thumbnail/', methods=['POST'])
 @login_required
@@ -367,7 +193,6 @@ def photo_thumbnail():
 
     return app.response_class(response="[]", mimetype='application/json')
 
-
 @app.route('/rest/gallery/', methods=['GET'])
 @login_required
 def gallery_index():
@@ -383,7 +208,7 @@ def gallery_index():
 
     limit = 5
     offset = (page_num - 1) * limit
-    galleries = find_gallery_all(offset=offset, limit=limit, search_query=search_query)
+    galleries = Gallery.find_all(offset=offset, limit=limit, search_query=search_query)
 
     response = []
     for gallery in galleries:
@@ -395,7 +220,7 @@ def gallery_index():
 @login_required
 def gallery_get(gallery_id):
     """ Get a gallery """
-    gallery = find_gallery_by_id(gallery_id)
+    gallery = Gallery.find_by_id(gallery_id)
     if not gallery:
         abort(404)
 
@@ -409,8 +234,7 @@ def gallery_post():
         abort(404)
 
     gallery = Gallery(name=request.form['name'])
-    db.session.add(gallery)
-    db.session.commit()
+    gallery.save()
 
     response = gallery.to_object()
 
@@ -424,7 +248,7 @@ def gallery_item(gallery_id):
     if current_user.role == "guest":
         abort(404)
 
-    gallery = find_gallery_by_id(gallery_id)
+    gallery = Gallery.find_by_id(gallery_id)
     if not gallery:
         abort(404)
 
@@ -435,8 +259,7 @@ def gallery_item(gallery_id):
         gallery.delete()
 
         # Delete the gallery from the database
-        db.session.delete(gallery)
-        db.session.commit()
+        gallery.delete()
         response = []
 
     return app.response_class(response=json.dumps(response), mimetype='application/json')
@@ -477,7 +300,7 @@ def photo_delete(gallery_id, photo_id):
     if current_user.role == "guest":
         abort(404)
 
-    gallery = find_gallery_by_id(gallery_id)
+    gallery = Gallery.find_by_id(gallery_id)
     if not gallery:
         abort(404)
