+++ app/forms/user.py
@@ -1,22 +1,8 @@
 from flask.ext.wtf import Form
 from wtforms.fields import TextField,\
 						   PasswordField,\
-						   BooleanField,\
-						   DateField,\
 						   SelectField
 from wtforms.validators import Required
-from datetime import datetime
-
-
-class LoginForm(Form):
-    username = TextField('Username', validators=[Required()])
-    password = PasswordField('Password', validators=[Required()])
-    remember = BooleanField('Remember Me', default=False)
-
-
-class GalleryForm(Form):
-    name = TextField('Name', validators=[Required()])
-    date = DateField('Date', validators=[Required()])
 
 
 class UserForm(Form):
