+++ app/views/helpers/filters.py
