+++ app/views/templates/gallery/index.html
