+++ app/views/templates/gallery/update.html
