+++ app/views/templates/users/update.html
@@ -4,7 +4,7 @@
 
 {% block body %}
 
-<form class="form form-horizontal col-sm-4 col-sm-offset-4" action="{{ form_action }}" method="post">
+<form class="form form-vertical col-sm-4 col-sm-offset-4" action="{{ form_action }}" method="post">
     {{ form.csrf_token }}
 
     {% if form.errors %}
