+++ package.json
@@ -11,8 +11,8 @@
     "grunt": "*",
     "grunt-cli": "*",
     "grunt-react": "*",
-    "grunt-contrib-uglify": "*",
     "grunt-contrib-less": "*",
+    "grunt-contrib-uglify": "*",
     "grunt-contrib-watch": "*",
     "react-tools": "*",
     "uglify-js": "*"
