+++ requirements.txt
@@ -1,13 +1,13 @@
 boto==2.38.0
-Flask-Compress==1.0.0
-Flask-Login==0.2.9
-Flask-Migrate==1.2.0
-Flask-Script==0.6.6
+Flask-Compress==1.2.1
+Flask-Login==0.2.11
+Flask-Migrate==1.5.0
+Flask-Script==2.0.5
 Flask-SQLAlchemy==2.0
-Flask-WTF==0.10.2
+Flask-WTF==0.12
 Flask==0.10.1
+MySQL-python==1.2.5
 PyReact==0.5.2
 pytz==2014.7
-tornado==4.0.2
+tornado==4.2.1
 urldecode==0.1
-MySQL-python==1.2.5
